requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
pydantic>=2.6.0
PyYAML>=6.0
networkx>=3.2
requests>=2.31.0
streamlit>=1.32.0
pytest>=8.0.0
